refactored test_utils => test_constants
